+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This page lists the changes to the IsaacScript framework.
 
 ## June 20th, 2022
 
+- Breaking changes:
+  - `DefaultMap` now takes the default value/function as the first argument and the initializer array as the second argument.
 - Added the following helper functions:
   - `isPrimitive`